requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
conda-pack @ file:///C:/b/abs_7b5jwr84ex/croot/conda-pack_1752674284442/work
contourpy==1.3.0
cycler==0.12.1
filelock==3.19.1
fonttools==4.60.1
idna==3.11
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.8.0.76
packaging==25.0
pillow==11.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
requests==2.32.5
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2024.8.30
torch @ file:///D:/PyTorch_Packages/torch-2.0.0%2Bcu118-cp39-cp39-win_amd64.whl#sha256=893f6dd205316a104b04b69877bc40ab9908428274920094c17b81396a8c985c
torchaudio==2.0.0
torchvision==0.15.1+cu118
typing_extensions==4.15.0
urllib3==2.5.0
zipp==3.23.0
